# contracts/saas/openapi.yaml
openapi: 3.1.0
...
info:
  title: Dockling hosted API contract
  version: 0.4.1
  description: Versioned SaaS design. Hosted Phases 1–4 are implemented for synthetic
    local acceptance; billing, public deployment and future aliases remain planned.
    Uploads/workbook apply use bounded raw bytes. AI uses separate consent and conservative
    reservations; production teardown recovery remains a deployment gate.
servers:
- url: /api/v1
security:
- SessionCookie: []
paths:
...
  /jobs/{job}:
    get:
      operationId: get_jobs_job
      summary: GET /jobs/{job}
      x-phase: 1
      x-feature-ids:
      - F02
      x-access-roles:
      - owner
      - editor
      - viewer
      - support_read
      x-workspace-scoped: true
      responses:
        '200':
          description: Successful authorized result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/PathJob'
    patch:
      operationId: patch_jobs_job
      summary: PATCH /jobs/{job}
      x-phase: 2
      x-feature-ids:
      - F01
      - F09
      - F26
      x-access-roles:
      - owner
      - editor
      x-workspace-scoped: true
      responses:
        '200':
          description: Successful authorized result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/PathJob'
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobUpdate'
...
  /jobs/{job}/rows:
    get:
      operationId: get_jobs_job_rows
      summary: GET /jobs/{job}/rows
      x-phase: 2
      x-feature-ids:
      - F11
      x-access-roles:
      - owner
      - editor
      - viewer
      - support_read
      x-workspace-scoped: true
      responses:
        '200':
          description: Successful authorized result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/PathJob'
      - $ref: '#/components/parameters/QueryLimit'
      - $ref: '#/components/parameters/QueryCursor'
      description: Read-only source-linked canonical rows at the current revision.
        Monetary values remain exact decimal strings; raw text and full account identity
        are excluded. Editing/export remains Phase 3.
  /jobs/{job}/review:
    post:
      operationId: post_jobs_job_review
      summary: POST /jobs/{job}/review
      x-phase: 3
      x-feature-ids:
      - F11
      x-access-roles:
      - owner
      - editor
      x-workspace-scoped: true
      responses:
        '200':
          description: Successful authorized result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/PathJob'
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCommand'
      x-implementation: planned; use documented hosted routes for implemented capability
    patch:
      summary: Edit
      operationId: hosted_patch__jobs_job_review
      parameters:
      - name: job
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Jid
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedEditCommand'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Edit Api V1 Jobs  Jid  Review Patch
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 3
      x-implementation: implemented
      x-workspace-scoped: true
      x-access-roles:
      - owner
      - editor
      x-feature-ids:
      - F11
      - F12
...
  /jobs/{job}/exports:
    post:
      summary: Prepare
      operationId: hosted_post__jobs_job_exports
      parameters:
      - name: job
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Jid
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedRevisionCommand'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Prepare Api V1 Jobs  Jid  Exports Post
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 3
      x-implementation: implemented
      x-workspace-scoped: true
      x-access-roles:
      - owner
      - editor
      x-feature-ids:
      - F13
      - F14
      - F15
      - F17
      - F18
      - F19
      - F20
    get:
      summary: Outputs
      operationId: hosted_get__jobs_job_exports
      parameters:
      - name: job
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Jid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Outputs Api V1 Jobs  Jid  Exports Get
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 3
      x-implementation: implemented
      x-workspace-scoped: true
      x-access-roles:
      - owner
      - editor
      - viewer
      x-feature-ids:
      - F13
      - F14
      - F15
      - F17
      - F18
      - F19
      - F20
...
  /jobs/{job}/close:
    post:
      summary: Close
      operationId: hosted_post__jobs_job_close
      parameters:
      - name: job
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Jid
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedRemovalCommand'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Close Api V1 Jobs  Jid  Close Post
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 3
      x-implementation: implemented
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F21
      - F22
...
  /connections:
    get:
      summary: Connections
      operationId: hosted_get__connections
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      - editor
      - viewer
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
    post:
      summary: Create
      operationId: hosted_post__connections
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedConnectionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedConnectionRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F24
      - F25
      - F26
      - F27
      parameters:
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
...
  /jobs/{job}/consent:
    get:
      operationId: get_jobs_job_consent
      summary: GET /jobs/{job}/consent
      x-phase: 4
      x-feature-ids:
      - F28
      x-access-roles:
      - owner
      - editor
      - viewer
      x-workspace-scoped: true
      responses:
        '200':
          description: Successful authorized result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consent'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/PathJob'
    post:
      summary: Consent
      operationId: hosted_post__jobs_job_consent
      parameters:
      - name: job
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Jid
      - $ref: '#/components/parameters/HeaderXCsrfToken'
      - $ref: '#/components/parameters/HeaderIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostedConsentCommand'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostedAiStatusRead'
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 4
      x-implementation: 'implemented: SaaS Phase 4; single API/worker local topology,
        synthetic acceptance only'
      x-workspace-scoped: true
      x-access-roles:
      - owner
      x-feature-ids:
      - F28
      - F26
      - F27
...
  /jobs/{job}/privacy:
    get:
      summary: Policy
      operationId: hosted_get__jobs_job_privacy
      parameters:
      - name: job
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Jid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Policy Api V1 Jobs  Jid  Privacy Get
        '422':
          $ref: '#/components/responses/Error422'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '413':
          $ref: '#/components/responses/Error413'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      x-phase: 3
      x-implementation: implemented
      x-workspace-scoped: true
      x-access-roles:
      - owner
      - editor
      - viewer
      x-feature-ids:
      - F21
      - F22
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: dockling_session
      description: Opaque server session; Secure, HttpOnly, SameSite. Not an OIDC
        access token or browser role.
    PaymentSignature:
      type: apiKey
      in: header
      name: Payment-Signature
      description: Placeholder for the selected provider signature scheme; validate
        raw provider bytes, timestamp and replay policy before processing.
  schemas:
...
    Job:
      type: object
      properties:
        id: *id001
        name: *id006
        status: *id007
        revision: *id004
        currency: *id008
        created_at: *id005
        deletion_state: *id009
        options:
          $ref: '#/components/schemas/JobOptions'
        content_digest:
          anyOf:
          - &id010
            type: string
            pattern: ^[0-9a-f]{64}$
          - type: 'null'
        export_revision:
          anyOf:
          - *id004
          - type: 'null'
        source_checked: *id003
        ai_source_checked: *id003
      required:
      - id
      - name
      - status
      - revision
      - currency
      - created_at
      - deletion_state
      - options
      - content_digest
      - export_revision
      - source_checked
      - ai_source_checked
      additionalProperties: false
...
  responses:
    Error400:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error401:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error403:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: Nonrevealing missing or unauthorized private resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error409:
      description: Revision/idempotency/state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error413:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error422:
      description: Safe structured error without private input echo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...